Added Support for Moving across multiple drives
In some cases (although rare) the source directory and target directory may be on different volumes/partitions/drives (as was with my case). This simple code that was taken off of stackoverflow.com and modified a tiny bit will move a folder from a drive to another drive with no problem.